bugfixing adding missing RunRequestScoped annotations, adding log output, and proper Long value comparison  refs #197

+++ backmeup-indexer-core/src/main/java/org/backmeup/index/core/model/IndexFragmentEntryStatus.java
@@ -105,7 +105,7 @@ public class IndexFragmentEntryStatus {
      * @return
      */
     public boolean isUserOwned() {
-        if (this.userID == this.ownerID) {
+        if (this.userID.longValue() == this.ownerID.longValue()) {
             return true;
         }
         return false;
